rearranged and renamed art page related files
