+++ styles/ArtCarouselCard.module.css
